Charge fractional credits for actual token and TTS usage (#105)
* Support fractional credits and character-priced ElevenLabs playback

* Verify exact financial fingerprints across the SQL migration

* Verify purchases and refunds retain fractional remainders

* Bound token settlement by its reservation and preserve large balance displays

---------

+++ Glosify/Models/Entities/AiCreditAccount.cs
@@ -3,12 +3,12 @@ namespace Glosify.Models.Entities;
 public sealed class AiCreditAccount
 {
     public string UserId { get; set; } = string.Empty;
-    public int BalanceCredits { get; set; }
-    public int ReservedCredits { get; set; }
+    public decimal BalanceCredits { get; set; }
+    public decimal ReservedCredits { get; set; }
     public DateTimeOffset? TrialGrantedAt { get; set; }
     public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow;
     public DateTimeOffset UpdatedAt { get; set; } = DateTimeOffset.UtcNow;
     public byte[] RowVersion { get; set; } = [];
 
-    public int AvailableCredits => BalanceCredits - ReservedCredits;
+    public decimal AvailableCredits => BalanceCredits - ReservedCredits;
 }
